tidy up flume, and move it to beside the db not inside it
